tabify, gitignore, add -Wextra to CFLAGS

+++ main.c
@@ -20,8 +20,8 @@ Vector2 window_size = {.x = INIT_WIDTH, .y = INIT_HEIGHT};
 void CustomLogCallback(int logLevel, const char *text, va_list args) {
 	FILE *out = stderr;
 	fprintf(out, "[%s] ", (logLevel == LOG_INFO) ? "INFO" :
-			      (logLevel == LOG_WARNING) ? "WARN" :
-			      (logLevel == LOG_ERROR) ? "ERR" : "DEBUG");
+		(logLevel == LOG_WARNING) ? "WARN" :
+		(logLevel == LOG_ERROR) ? "ERR" : "DEBUG");
 
 	vfprintf(out, text, args);
 	fprintf(out, "\n");
@@ -33,9 +33,9 @@ int main() {
 	SetWindowState(FLAG_WINDOW_RESIZABLE);
 	EnableEventWaiting();
 
-        struct lines *lines;
-        struct line *line;
-        lines_init(&lines);
+	struct lines *lines;
+	struct line *line;
+	lines_init(&lines);
 	line_init(&line);
 
 	while (!WindowShouldClose()) {
@@ -66,16 +66,16 @@ int main() {
 			ClearBackground(RAYWHITE);
 			/* DrawText(lineh->vec->data, */
 			/*	 (window_size.x - text_width) / 2, 0, */
-                        /*	 font_size, BLACK); */
-                        int i;
-                        for (i = 0; i < lines->arr->size; i++) {
+			/*	 font_size, BLACK); */
+			size_t i;
+			for (i = 0; i < lines->arr->size; i++) {
 				struct line *p = VEC_LINE_get(lines->arr, i);
 				DrawText(p->vec->data, 10, 10 + font_size * i, font_size, BLACK);
-                        }
-                        DrawText(line->vec->data, 10, 10 + font_size * i, font_size, BLACK);
+			}
+			DrawText(line->vec->data, 10, 10 + font_size * i, font_size, BLACK);
 			EndDrawing();
 		}
 	}
-        CloseWindow();
-        lines_free(lines);
+	CloseWindow();
+	lines_free(lines);
 }
